Refactor main func to the `run` subcommand in Cobra
